Palettes, bobines et kilos se repondent sur tous les ecrans de saisie
Le mouvement savait deja convertir les trois. La reception, le bon de
commande, le transfert et la facture d'import ne le savaient pas : il
fallait convertir de tete, ou saisir les trois separement en esperant
qu'elles concordent.

Desormais on saisit celui qu'on a sous les yeux — au camion des palettes,
sur la machine des bobines, a la bascule des kilos — et les deux autres
suivent les PARAMETRES DE LA REFERENCE, jamais un facteur invente par
l'ecran. Un interrupteur par ligne detache le calcul : une palette
entamee reste une palette a manutentionner, et c'est l'operateur qui le
sait.

Ce que chaque ecran gagne :

- RECEPTION : une colonne « Pal. / Bob. » dans la grille, le triplet
  complet dans l'onglet Quantite, et le selecteur d'unite n'offre que ce
  que la reference sait convertir. Les comparaisons (ecart au commande,
  ecart au bon de livraison, kg par colis, total) se font maintenant EN
  KILOS : elles prenaient la quantite saisie pour des kilos, ce qui etait
  faux des que l'unite ne l'etait pas.
- BON DE COMMANDE : quantite, unite, palettes et bobines par ligne. Le
  prix reste saisi au KILO — c'est ainsi que les offres se comparent — et
  se convertit vers l'unite commandee a l'envoi.
- TRANSFERT : ses colonnes bobines et palettes existaient deja, mais se
  saisissaient a la main, sans lien avec la quantite. Sa conversion en kg
  repliait sur un facteur NUL quand la reference ne portait pas son poids
  de bobine : dix bobines pesaient alors zero kilo, le total etait faux
  et rien ne le disait. Elle refuse desormais, et le dit.
- FACTURE D'IMPORT : poids net, bobines et palettes se repondent. Une
  ligne deja enregistree arrive DETACHEE, a dessein : ses comptes
  viennent du document du fournisseur, pas d'un calcul.

Partout, une unite que la reference ne sait pas convertir est signalee
pendant la saisie plutot qu'a l'enregistrement : le serveur la refuse
(R01, aucun repli sur un facteur de 1), et decouvrir le refus apres avoir
tout saisi fait perdre le formulaire.

Au passage, un defaut qui rendait l'ecran de creation d'un bon de
commande INUTILISABLE : `ROUND(x / $3, 6)` ou `$3` est lie depuis un
`f64`. La division rend un `double precision`, et PostgreSQL n'a pas de
`round(double precision, integer)`. La requete entiere echouait en 500 —
l'ecran n'affichait donc aucune reference a commander, sans un mot
expliquant pourquoi. Le taux est maintenant coule en `numeric`.

Verifie au navigateur sur le serveur d'essai : 30 controles, aucun echec.
Les trois sens de conversion sur les quatre ecrans, le detachement, et le
refus silencieux d'une reference sans conditionnement.

+++ backend/src/routes/stock.rs
@@ -1660,7 +1660,13 @@ pub async fn references_commandables(
                                 AND to_char(current_date, 'YYYY-MM-DD') >= substr(t.date_debut, 1, 10)
                                 AND (t.date_fin IS NULL OR to_char(current_date, 'YYYY-MM-DD') < substr(t.date_fin, 1, 10))
                               ORDER BY t.date_debut DESC LIMIT 1), 1.0))
-                      / $3, 6) AS prix_suggere_devise,
+                      -- LE TAUX ARRIVE EN `float8` (il est lie depuis un `f64`).
+                      -- Sans ce cast, la division rend un `double precision`,
+                      -- et PostgreSQL n'a pas de `round(double precision, int)` :
+                      -- la requete entiere echouait en 500, donc l'ecran de
+                      -- creation d'un bon n'affichait AUCUNE reference — sans
+                      -- un mot expliquant pourquoi.
+                      / $3::numeric, 6) AS prix_suggere_devise,
                 CASE WHEN pa.prix_estime_mad IS NOT NULL THEN pa.source_prix
                      WHEN r.cmup_mad IS NOT NULL THEN 'CMUP'
                      ELSE 'CATALOGUE' END AS source_prix,
